extract network concern
