+++ packages/unraid-api-plugin-connect/src/network/network.resolver.ts
@@ -9,7 +9,7 @@ import {
 } from '@unraid/shared/use-permissions.directive.js';
 
 import { Network } from '../model/connect.model.js';
-import { UrlResolverService } from '../service/url-resolver.service.js';
+import { UrlResolverService } from './url-resolver.service.js';
 
 @Resolver(() => Network)
 export class NetworkResolver {
