+++ packages/unraid-api-plugin-connect/src/remote-access/remote-access.module.ts
@@ -4,16 +4,16 @@ import { WanAccessEventHandler } from '../event-handler/wan-access.events.js';
 import { DynamicRemoteAccessService } from './dynamic-remote-access.service.js';
 import { StaticRemoteAccessService } from './static-remote-access.service.js';
 import { UpnpRemoteAccessService } from './upnp-remote-access.service.js';
-import { SystemModule } from '../module/system.module.js';
+import { NetworkModule } from '../network/network.module.js';
 
 @Module({
-    imports: [SystemModule],
+    imports: [NetworkModule],
     providers: [
         DynamicRemoteAccessService,
         StaticRemoteAccessService,
         UpnpRemoteAccessService,
         WanAccessEventHandler,
     ],
-    exports: [DynamicRemoteAccessService, SystemModule],
+    exports: [DynamicRemoteAccessService, NetworkModule],
 })
 export class RemoteAccessModule {}
