+++ packages/unraid-api-plugin-connect/src/service/connect-settings.service.ts
@@ -26,7 +26,7 @@ import { ConfigType, MyServersConfig } from '../config/connect.config.js';
 import { EVENTS } from '../helper/nest-tokens.js';
 import { DynamicRemoteAccessType, WAN_ACCESS_TYPE, WAN_FORWARD_TYPE } from '../model/connect.model.js';
 import { DynamicRemoteAccessService } from '../remote-access/dynamic-remote-access.service.js';
-import { NetworkService } from './network.service.js';
+import { NetworkService } from '../network/network.service.js';
 
 declare module '@unraid/shared/services/user-settings.js' {
     interface UserSettings {
